+++ resource-server/src/main/java/org/osiam/resources/exception/ResourceNotFoundException.java
@@ -21,14 +21,9 @@
  * SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
 
-package org.osiam.resources.exceptions;
+package org.osiam.resources.exception;
 
-import org.springframework.http.HttpStatus;
-import org.springframework.web.bind.annotation.ResponseStatus;
-
-@ResponseStatus(value = HttpStatus.NOT_FOUND)
 public class ResourceNotFoundException extends OsiamException {
-    private static final long serialVersionUID = -690283630029161964L;
 
     public ResourceNotFoundException(String s) {
         super(s);