+++ resource-server/src/main/java/org/osiam/resources/provisioning/update/UserUpdater.java
@@ -26,8 +26,8 @@ package org.osiam.resources.provisioning.update;
 import java.util.HashSet;
 import java.util.Set;
 
-import org.osiam.resources.exceptions.OsiamException;
-import org.osiam.resources.exceptions.ResourceExistsException;
+import org.osiam.resources.exception.OsiamException;
+import org.osiam.resources.exception.ResourceExistsException;
 import org.osiam.resources.scim.User;
 import org.osiam.storage.dao.UserDao;
 import org.osiam.storage.entities.UserEntity;