+++ resource-server/src/test/groovy/org/osiam/resources/exception/ResourceNotFoundExceptionSpec.groovy
@@ -21,31 +21,28 @@
  * SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
 
-package org.osiam.resources.exceptions;
+package org.osiam.resources.exception
 
-import java.util.regex.Matcher;
-import java.util.regex.Pattern;
+import spock.lang.Specification
 
-public class JsonPropertyMessageTransformer implements ErrorMessageTransformer {
+class ResourceNotFoundExceptionSpec extends Specification {
 
-    // should only get Unrecognized field and name of the field e.q.:
-    // Unrecognized field "extId" (Class org.osiam.resources.scim.User), not marked as ignorable
-    // at [Source: java.io.StringReader@1e41ac03; line: 1, column: 11] (through reference chain:
-    // org.osiam.resources.scim.User["extId"])
-    // will be transformed to
-    // Unrecognized field "extId"
-    private static final Pattern PATTERN = Pattern.compile("(Unrecognized field \"\\w+\").*",
-            Pattern.CASE_INSENSITIVE | Pattern.DOTALL | Pattern.MULTILINE);
+    def "should contain given message"() {
+        when:
+        def rnfe = new ResourceNotFoundException("haha")
+        then:
+        rnfe.message == "haha"
+    }
 
-    @Override
-    public String transform(String message) {
-        if (message == null) {
-            return null;
-        }
-        Matcher matcher = PATTERN.matcher(message);
-        if (matcher.matches()) {
-            return matcher.group(1);
-        }
-        return message;
+    def "should contain constructor with message and cause"() {
+        given:
+        def cause = Mock(Throwable)
+
+        when:
+        def rnfe = new ResourceNotFoundException("haha", cause)
+
+        then:
+        rnfe.message == "haha"
+        rnfe.getCause() == cause
     }
 }